Move iframe block to its own file
