+++ cms/datavis/blocks/__init__.py
@@ -2,10 +2,10 @@ from .charts import (
     AreaChartBlock,
     BarColumnChartBlock,
     BarColumnConfidenceIntervalChartBlock,
-    IframeBlock,
     LineChartBlock,
     ScatterPlotBlock,
 )
+from .iframe import IframeBlock
 
 __all__ = [
     "AreaChartBlock",
